+++ modules/doofinder-sync/includes/class-doofinder-core.php
@@ -307,26 +307,12 @@ class Doofinder_Core
      */
     public static function add_price_structure_fix(): void
     {
-        ?>
-        <script>
-            jQuery(document).ready(function ($) {
-                $('.price').each(function () {
-                    var $price = $(this);
-                    var $del = $price.find('del');
-                    if ($del.find('ins').length > 0) {
-                        var originalPriceHTML = $del.find('ins .woocommerce-Price-amount').parent().html();
-                        var currentPriceHTML = $price.find('del + ins .woocommerce-Price-amount').parent().html();
-                        var originalPriceText = $del.find('ins .woocommerce-Price-amount').text();
-                        var currentPriceText = $price.find('del + ins .woocommerce-Price-amount').text();
-                        var newHtml = '<del aria-hidden="true">' + originalPriceHTML + '</del> ';
-                        newHtml += '<span class="screen-reader-text">Original price was: ' + originalPriceText + '.</span>';
-                        newHtml += '<ins>' + currentPriceHTML + '</ins>';
-                        newHtml += '<span class="screen-reader-text">Current price is: ' + currentPriceText + '.</span>';
-                        $price.html(newHtml);
-                    }
-                });
-            });
-        </script>
-        <?php
+        wp_enqueue_script(
+            'doofinder-price-fix',
+            plugin_dir_url(dirname(__FILE__)) . 'assets/js/price-structure-fix.js',
+            ['jquery'],
+            FFLA_VERSION,
+            true
+        );
     }
 }